+++ lib/duffer.dart
@@ -6,6 +6,7 @@ export 'src/bytebuf_base.dart';
 export 'src/extensions.dart';
 export 'src/pooled.dart';
 export 'src/serialization/pickle.dart';
+export 'src/serialization/polymorphic.dart';
 export 'src/unpooled.dart';
 export 'src/utils/constants.dart';
 export 'src/utils/migration_utils.dart';