+++ lib/src/extensions.dart
@@ -2,12 +2,13 @@ import 'dart:convert';
 import 'dart:typed_data';
 
 import 'bytebuf_base.dart';
-import 'utils/constants.dart';
 import 'impl/list_buffer.dart';
+import 'utils/constants.dart';
 
-part 'extensions/get_set.dart';
-part 'extensions/read_write.dart';
-part 'extensions/logging.dart';
-part 'extensions/qol.dart';
 part 'extensions/encoding.dart';
+part 'extensions/get_set.dart';
+part 'extensions/logging.dart';
 part 'extensions/migrations.dart';
+part 'extensions/qol.dart';
+part 'extensions/read_write.dart';
+part 'extensions/shift.dart';