+++ lib/src/extensions/shift.dart
@@ -1,0 +1,6 @@
+part of '../extensions.dart';
+
+extension ShiftExtension on ByteBuf {
+  //TODO: Implement Shifting
+
+}